requirements.txt
asgiref==3.3.1
beautifulsoup4==4.10.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-crispy-forms==1.10.0
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-filter==21.1
django-phone-field==1.8.1
django-search-admin-autocomplete==0.2.0
djangorestframework==3.12.2
dnspython==2.1.0
drf-yasg==1.20.0
emoji==1.5.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
MouseInfo==0.1.3
msgpack==1.0.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pathlib==1.0.1
Pillow==8.1.0
proto-plus==1.18.1
protobuf==3.17.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
pydantic==1.8.2
pyfiglet==0.8.post1
PyGetWindow==0.0.9
PyJWT==2.3.0
pymongo==3.12.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.1
python3-xlib==0.15
pytweening==1.0.4
pytz==2020.5
pywhatkit==5.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
twilio==7.5.0
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.5
verlat==0.1.0.post1
wappdriver==0.6.6
whitenoise==5.2.0
wikipedia==1.4.0
